+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes of this package are documented in this file using the [Keep 
 
 [Keep a Changelog]: https://keepachangelog.com/
 
+## Unreleased
+
+### Added
+
+* [wiz-pkg] Add `wiz-pkg` function for package installation.
+
 ## [0.0.3]
 
 ### Added